refactor: 整合ZookeeperRegistry
