+++ motan-demo/motan-demo-client/src/main/resources/motan_demo_client_commandRegistry.xml
@@ -21,7 +21,7 @@
        http://api.weibo.com/schema/motan http://api.weibo.com/schema/motan.xsd">
 
     <!-- 注册中心配置 使用不同注册中心需要依赖对应的jar包。-->
-    <motan:registry regProtocol="commandZookeeper" name="registry" address="127.0.0.1:2181"/>
+    <motan:registry regProtocol="zookeeper" name="registry" address="127.0.0.1:2181"/>
 
     <!-- motan协议配置 -->
     <motan:protocol default="true" name="motan" haStrategy="failfast" loadbalance="configurableWeight"/>
