+++ motan-demo/motan-demo-server/src/main/java/com/weibo/motan/demo/server/MotanApiExportDemo.java
@@ -48,13 +48,6 @@ public class MotanApiExportDemo {
         zookeeperRegistry.setAddress("127.0.0.1:2181");
         motanDemoService.setRegistry(zookeeperRegistry);
 
-        // 配置支持指令的ZooKeeper注册中心
-        // RegistryConfig commandZKRegistry = new RegistryConfig();
-        // commandZKRegistry.setRegProtocol("commandZookeeper");
-        // commandZKRegistry.setAddress("127.0.0.1:2181");
-        // motanDemoService.setRegistry(commandZKRegistry);
-
-
         // 配置RPC协议
         ProtocolConfig protocol = new ProtocolConfig();
         protocol.setId("motan");
