+++ motan-demo/motan-demo-server/src/main/resources/motan_demo_server_commandRegistry.xml
@@ -25,7 +25,7 @@
     <bean id="motanDemoServiceImpl" class="com.weibo.motan.demo.server.MotanDemoServiceImpl"/>
 
     <!-- 注册中心配置 使用不同注册中心需要依赖对应的jar包。如果不使用注册中心，可以把check属性改为false，忽略注册失败。-->
-    <motan:registry regProtocol="commandZookeeper" name="registry" address="127.0.0.1:2181"/>
+    <motan:registry regProtocol="zookeeper" name="registry" address="127.0.0.1:2181"/>
 
     <!-- 协议配置。为防止多个业务配置冲突，推荐使用id表示具体协议。-->
     <motan:protocol id="demoMotan" default="true" name="motan"/>
